requirements.txt
pandas>=2.0
numpy>=1.26
pydantic>=2.6
regex>=2023.10.3
rapidfuzz>=3.5
scikit-learn>=1.4
rank-bm25>=0.2.2
jinja2>=3.1
# dev
pre-commit>=3.7
black>=24.4
isort>=5.13
flake8>=7.0
PyYAML>=6.0
jsonschema>=4.22
